postprocessing/requirements.txt
matplotlib
kafka-python
nexradaws
numpy==1.16.1
arm-pyart
netCDF4
scipy
pandas
